+++ symja_android_library/matheclipse-core/src/main/java/org/matheclipse/core/builtin/StatisticsFunctions.java
@@ -1801,7 +1801,7 @@ public class StatisticsFunctions {
 							return F.Times(F.CN1, num1.setAtClone(i, F.Times(factor, arg1.get(i))), F.Conjugate(arg2.get(i)));
 						}
 					});
-			return F.Divide(v1, F.integer((arg1.argSize()) * (arg1.size() - 2)));
+			return F.Divide(v1, F.integer(((long)arg1.argSize()) * (((long)arg1.size()) - 2L)));
 		}
 
 		@Override
